@@ -1,6 +1,6 @@
 {
   "name": "basic-to-passport-auth-http-proxy",
-  "version": "0.2.0-beta.0",
+  "version": "0.2.1",
   "description": "HTTP proxy server that can access resources which use the Passport SSI Version 1.4 Protocol for authorization with credentials supplied by Basic HTTP authentication.",
   "keywords": [
     "http",
@@ -8,6 +8,7 @@
     "Basic authentication",
     "Passport 1.4 authentication"
   ],
+  "type": "module",
   "license": "MIT",
   "author": {
     "name": "Stefan Kleeschulte",
@@ -17,7 +18,7 @@
     "type": "git",
     "url": "https://github.com/skleeschulte/basic-to-passport-auth-http-proxy.git"
   },
-  "main": "lib/server.js",
+  "main": "lib/main.js",
   "scripts": {
     "test:local": "mocha test/local/index.js",
     "test:onedrive_litmus": "mocha test/onedrive_litmus/index.js",
@@ -28,19 +29,17 @@
     "node": ">=10.16.0"
   },
   "dependencies": {
-    "axios": "1.6.7",
-    "debug": "4.3.4",
-    "get-request-url": "0.1.1",
-    "http-proxy": "1.18.1",
-    "supports-color": "9.4.0",
-    "tough-cookie": "4.1.3"
+    "axios": "^1.7.1",
+    "debug": "^4.3.4",
+    "get-request-url": "^0.1.1",
+    "http-proxy": "^1.18.1",
+    "supports-color": "^9.4.0",
+    "tough-cookie": "^4.1.4"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "dotenv": "^10.0.0",
-    "eslint": "^5.3.0",
-    "eslint-config-airbnb-base": "^13.2.0",
-    "eslint-plugin-import": "^2.17.2",
-    "mocha": "^6.1.4"
+    "chai": "^5.1.1",
+    "dotenv": "^16.4.5",
+    "eslint": "^9.3.0",
+    "mocha": "^10.4.0"
   }
 }
