@@ -1,9 +1,9 @@
-const parsePassportParameters = require('./parsePassportParameters');
+import { parsePassportParameters } from './parsePassportParameters.js';
 
 /**
  * Class for parsing HTTP auth headers (e.g. WWW-Authenticate or Authentication-Info).
  */
-class AuthHeader {
+export class AuthHeader {
     /**
      * Parse Basic auth header param.
      *
@@ -45,5 +45,3 @@ class AuthHeader {
         }
     }
 }
-
-module.exports = AuthHeader;
