@@ -1,4 +1,4 @@
-const tough = require('tough-cookie');
+import { CookieJar, Cookie } from 'tough-cookie';
 
 /**
  * The cookieJar from tough-cookie separates cookies depending on the url. For Passport authentication, the same cookies
@@ -13,14 +13,14 @@ const PASSPORT_COOKIE_URL = 'https://localpassportcookiesstore-pl8q2i8k0j69autii
 /**
  * Holds login session specific data.
  */
-class LoginSession {
+export class LoginSession {
     /**
      * Initialize a new LoginSession instance.
      */
     constructor() {
         this.passwordHash = null;
         // this.cookieJar = new tough.CookieJar({ rejectPublicSuffixes: false }); // leads to sync functions not working
-        this.cookieJar = new tough.CookieJar();
+        this.cookieJar = new CookieJar();
         this.lastAccessed = Date.now();
         this.passportCookieDomain = new URL(PASSPORT_COOKIE_URL).hostname;
 
@@ -36,7 +36,7 @@ class LoginSession {
      */
     setPassportCookies(cookieStrings) {
         cookieStrings.forEach((cookieString) => {
-            const cookie = tough.Cookie.parse(cookieString);
+            const cookie = Cookie.parse(cookieString);
 
             if (cookie.domain) {
                 cookie.domain = this.passportCookieDomain;
@@ -62,5 +62,3 @@ class LoginSession {
         this.cookieJar.removeAllCookiesSync();
     }
 }
-
-module.exports = LoginSession;
