@@ -1,4 +1,4 @@
-const crypto = require('crypto');
+import { createHash } from 'crypto';
 
 /**
  * Session time to live (TTL) in milliseconds.
@@ -9,7 +9,7 @@ const LOGIN_SESSION_TTL = 15 * 60 * 1000; // 15 minutes
 /**
  * Store for LoginSessions. LoginSessions are identified by username and protected by password provided by Basic auth.
  */
-class LoginSessionStore {
+export class LoginSessionStore {
     /**
      * Passport 1.4 usernames must be email names:
      * https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-pass/bd678e19-bc40-478e-ac2e-dbd2c654605d
@@ -29,7 +29,7 @@ class LoginSessionStore {
      * @returns {string}  SHA1 password hash.
      */
     static hashPassword(password) {
-        return crypto.createHash('sha1').update(password).digest('base64');
+        return createHash('sha1').update(password).digest('base64');
     }
 
     /**
@@ -93,5 +93,3 @@ class LoginSessionStore {
         this.store.delete(LoginSessionStore.normalizeUsername(username));
     }
 }
-
-module.exports = LoginSessionStore;
