@@ -1,8 +1,8 @@
-const axios = require('axios');
-const AuthHeader = require('./AuthHeader');
-const PassportError = require('./PassportError');
-const parsePassportParameters = require('./parsePassportParameters');
-const log = require('./log');
+import axios from 'axios';
+import { AuthHeader } from './AuthHeader.js';
+import { PassportError } from './PassportError.js';
+import { parsePassportParameters } from './parsePassportParameters.js';
+import { log } from './log.js';
 
 /**
  * Configuration server URL.
@@ -14,7 +14,7 @@ const CONFIGURATION_SERVER_URL = process.env.CONFIGURATION_SERVER_URL || 'https:
 /**
  * Class for handling Passport 1.4 authentication.
  */
-class PassportHandler {
+export class PassportHandler {
     /**
      * RFC3986 compliant encodeURIComponent function. It does encode commas (',').
      * Source: https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/encodeURIComponent
@@ -342,7 +342,7 @@ class PassportHandler {
             // and https://login.live.com/login2.srf return srealm and DARealm with different letter cases. So do a best
             // effort approach and compare lower case values.
             if (authenticateHeader.parameters['srealm'].toLowerCase()
-                    !== this.passportSession.passportConfigurationData['DARealm'].toLowerCase()) {
+                !== this.passportSession.passportConfigurationData['DARealm'].toLowerCase()) {
                 errorMessage = 'Authentication server realm does not equal realm in passport configuration data';
             }
 
@@ -400,5 +400,3 @@ class PassportHandler {
         throw new PassportError('Did not receive Set Token Message after First Authenticated Request Message');
     }
 }
-
-module.exports = PassportHandler;
