@@ -1,4 +1,4 @@
-const log = require('./log');
+import { log } from './log.js';
 
 /**
  * If the request body data exceeds this size, caching will be aborted.
@@ -9,7 +9,7 @@ const MAX_REQUEST_BODY_CACHE_SIZE = 2 * 1024; // 2 KiB
 /**
  * Class for caching client request bodies, so requests with body data can be repeated after successful authentication.
  */
-class RequestBodyCache {
+export class RequestBodyCache {
     /**
      * Returns the maximum cache size.
      *
@@ -110,5 +110,3 @@ class RequestBodyCache {
         });
     }
 }
-
-module.exports = RequestBodyCache;
