@@ -2,7 +2,7 @@
  * http-proxy's web outgoing pipeline is not exported, but grabbing it directly works:
  * @type {{}}
  */
-const httpProxyWebOutgoing = require('http-proxy/lib/http-proxy/passes/web-outgoing');
+import httpProxyWebOutgoing from 'http-proxy/lib/http-proxy/passes/web-outgoing.js';
 
 /**
  * Crate an array from the object that defines the outgoing passes.
@@ -21,7 +21,7 @@ const httpProxyWebOutgoingPasses = Object.keys(httpProxyWebOutgoing).map(pass =>
  * @param {http.ServerResponse}   res       HTTP response to the client
  * @param {object}                options   Proxy server options
  */
-function handleResponseWithHttpProxy(proxyRes, req, res, options) {
+export function handleResponseWithHttpProxy(proxyRes, req, res, options) {
     // Adapted from:
     // https://github.com/http-party/node-http-proxy/blob/1.17.0/lib/http-proxy/passes/web-incoming.js#L173-L177
     if (!res.headersSent) {
@@ -39,4 +39,3 @@ function handleResponseWithHttpProxy(proxyRes, req, res, options) {
     }
 }
 
-module.exports = handleResponseWithHttpProxy;
