@@ -1,11 +1,11 @@
-const debug = require('debug');
+import debug from 'debug';
 
 if (!process.env.DEBUG) {
     debug.enable('proxy:info,proxy:error');
 }
 
 // info (-> STDOUT)
-const log = debug('proxy:info');
+export const log = debug('proxy:info');
 log.log = console.log.bind(console); // eslint-disable-line no-console
 
 // error (-> STDERR)
@@ -20,4 +20,3 @@ log.t = debug('proxy:trace');
 log.t.log = console.log.bind(console); // eslint-disable-line no-console
 log.t.enabled = debug.enabled('proxy:trace');
 
-module.exports = log;
