@@ -1,4 +1,4 @@
-const log = require('./log');
+import { log } from './log.js';
 
 /**
  * Get proxy target from env.
@@ -19,7 +19,7 @@ try {
  * http-proxy configuration object
  * @type {{}}
  */
-module.exports = {
+export const proxyConfig = {
     target,
     changeOrigin: true,           // rewrite Origin HTTP header
     preserveHeaderKeyCase: true,  // preserve letter case of HTTP response headers
