@@ -1,7 +1,7 @@
 /**
  * Errors thrown by the PassportHandler.
  */
-class PassportError extends Error {
+export class PassportError extends Error {
     /**
      * Initialize a new PassportError.
      *
@@ -18,5 +18,3 @@ class PassportError extends Error {
         this.body = body;
     }
 }
-
-module.exports = PassportError;
