@@ -2,7 +2,7 @@
  * Holds Passport specific data with the lifetime of one client request/response cycle.
  * https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-pass/11a9752a-67c6-4fef-86ee-920e3f88f2c1
  */
-class PassportSession {
+export class PassportSession {
     /**
      * Initialize a new PassportSession.
      */
@@ -15,5 +15,3 @@ class PassportSession {
         this.signInAttempts = 0;
     }
 }
-
-module.exports = PassportSession;
