@@ -4,7 +4,7 @@
  * @param {string}  parameterString  String containing Passport parameters
  * @returns {{}} Object with parsed parameters.
  */
-function parsePassportParameters(parameterString) {
+export function parsePassportParameters(parameterString) {
     const elements = parameterString.split(',');
 
     return elements.reduce((parameters, element) => {
@@ -25,5 +25,3 @@ function parsePassportParameters(parameterString) {
         return parameters;
     }, {});
 }
-
-module.exports = parsePassportParameters;
